+++ myDate.js
@@ -1,0 +1,18 @@
+export function formatDate(date) {
+
+  let day = date.getDate();
+  day = day < 10 ? `0` + day : day;
+  let month = date.getMonth(date.setMonth(9));
+  month = month < 10 ? `0` + month : month;
+  let year = date.getFullYear();
+  year = year < 10 ? year = `0` + year : year;
+  year = year.toString().slice(-2);
+  let hours = date.getHours();
+  hours = hours < 10 ? hours = `0` + hours : hours;
+  let minutes = date.getMinutes();
+  minutes = minutes < 10 ? minutes = `0` + minutes : minutes;
+
+  let fullDate = `${day}:${month}.${year} ${hours}:${minutes}`;
+
+  return fullDate;
+}
